+++ Sources/BraintreeCore/URL+IsPayPal.swift
@@ -4,6 +4,6 @@ extension URL {
     
     /// Used to determine if a the URL is a paypal.com domain in order to format API requests accordingly
     var isPayPalURL: Bool {
-        return self.absoluteString.contains("api-m.paypal.com")
+        absoluteString.contains("api-m.paypal.com")
     }
 }